refactor tests  to use package scope setup and teardown
