+++ tests/helpers/common.py
@@ -18,3 +18,44 @@ def wait_for_line(fname, predicate, timeout=10, line_count=1):
             if times >= line_count:
                 f.close()
                 return True
+
+
+def assert_jsonrpc20_result(actual, expected):
+    assert actual['jsonrpc'] == '2.0'
+    assert actual['id'] == expected['id']
+    assert 'result' in actual
+    result = actual['result']
+    assert isinstance(result,  list)
+    assert 'error' not in actual
+    return result
+
+
+def equal(d1, d2, path=[]):
+    if isinstance(d1, dict):
+        if isinstance(d2, dict):
+            d1_keys = set(d1.keys())
+            d2_keys = set(d2.keys())
+            if len(d1_keys.difference(d2_keys)) > 0:
+                return [False, path]
+            for key in d1_keys:
+                path.append(key)
+                if not equal(d1[key], d2[key], path):
+                    return [False, path]
+                path.pop()
+
+            return [True, path]
+        else:
+            return [False, path]
+    elif isinstance(d1, list):
+        if isinstance(d2, list):
+            if len(d1) != len(d2):
+                return [False, path]
+            i = 0
+            for d1value, d2value in zip(d1, d2):
+                i += 1
+                path.append(i)
+                if not equal(d1value, d2value, path):
+                    return [False, path]
+            return [True, path]
+    else:
+        return [d2 == d1, path]
